+++ dvfs/jni/IOStuff.h
@@ -11,7 +11,7 @@
 #define IOSTUFF_H_
 
 void writeStringToFile(const char * filename, const char * value);
-void writeValueToFile(const char * filename, float value);
+void writeValueToFile(const char * filename, long value);
 float getValueFromFile(const char * filename);
 void getStringFromFile(const char * filename, char * buffer, int buffSize);
 void getStringFromFileByCat(const char * filename, char * buffer, int buffSize);
